Integrate waffle into contract:compile command
And go for native compilation
